Moved clock correction code into clockcorr.c/clockcorr.h file
